refactor(IT-Cotato#186): 게시글 익명 번호(댓글) 타입 수정

+++ src/main/java/com/cotato/kampus/domain/comment/application/CommentAppender.java
@@ -22,7 +22,7 @@ public class CommentAppender {
 	private final ApiUserResolver apiUserResolver;
 
 	@Transactional
-	public Long append(Long postId, String content, Long anonymousNumber, Long parentId, Long targetId){
+	public Long append(Long postId, String content, Integer anonymousNumber, Long parentId, Long targetId){
 
 		Long userId = apiUserResolver.getCurrentUserId();
